Added: Vagrantfile, README and gitignore
This commit adds a Vagrantfile to deliver a seamless experience while
running a virtual machine for the assignment.

In the process of spinning up the virutal machine, a shell script
(currently inline in the Vagrantfile) is run to install the necessary
dependencies to work on this project.

This commit also adds the .vagrant directory for the project, which
should be merged with #4 if it is merged.

Basics on how to install and use Vagrant are included in a new README.